+++ test/CoberturaConverter.Core.Tests/TestFile.cs
@@ -10,6 +10,7 @@
         DanglCommonOpenCover,
         LightQueryCobertura,
         LightQueryDotCover,
-        LightQueryOpenCover
+        LightQueryOpenCover,
+        NestedClassesDotCover
     }
 }