+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 
 All notable changes to **CoberturaConverter** are documented here.
 
+## v1.0.1
+- Support nested types when converting from dotCover
+- Exclude types without lines when converting from dotCover to Cobertura
+
 ## v1.0.0
 - Initial Release